backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
pyyaml==6.0
anthropic==0.7.0
python-telegram-bot==20.2
ccxt==3.0.0
requests==2.31.0
